+++ backend/src/models/product.ts
@@ -1,17 +1,6 @@
 import mongoose from 'mongoose';
 
-export type TProductImage = {
-    fileName: string;
-    originName: string;
-}
-
-export type TProduct = {
-    title: string;
-    image: TProductImage;
-    category: string;
-    description: string;
-    price: number;
-}
+import { TProduct, TProductImage  } from '../utils';
 
 const productShema = new mongoose.Schema<TProduct>({
     title: {
