+++ backend/src/utils/index.ts
@@ -1,2 +1,3 @@
 export { logger } from './logger/logger';
-
+export { productRouterValidation } from './validation';
+export { TProduct, TProductImage  } from './types';
