+++ backend/src/utils/types/index.ts
@@ -1,0 +1,2 @@
+export { type TProductImage, type TProduct  } from './product';
+
