+++ backend/src/utils/validation/index.ts
@@ -1,0 +1,2 @@
+export { productShema, productRouterValidation } from './product.validation';
+
